Splitter moved to core
