+++ documentation/src/main/cheatsheet/cheatsheets.adoc
@@ -146,6 +146,27 @@ Set the root folder of the repository on file system.
  If catalogue equals empty string a verticle will look for the files in classpath
 +++
 |===
+[[FragmentSplitterOptions]]
+== FragmentSplitterOptions
+
+++++
+ Describes FragmentSplitter Knot configuration
+++++
+'''
+
+[cols=">25%,^25%,50%"]
+[frame="topbot"]
+|===
+^|Name | Type ^| Description
+|[[address]]`address`|`String`|
++++
+Set the EB address of the verticle
++++
+|[[snippetTagName]]`snippetTagName`|`String`|
++++
+Sets a Knot.x snippet HTML tag name. Default is 'script'
++++
+|===
 [[HttpRepositoryOptions]]
 == HttpRepositoryOptions
