+++ documentation/src/main/wiki/KnotxDeployment.md
@@ -73,7 +73,7 @@ As mentioned above, the `knotx-starter.json` is the main configuration file desc
     "server=io.knotx.server.KnotxServerVerticle",
     "httpRepo=io.knotx.repository.http.HttpRepositoryConnectorVerticle",
     "fsRepo=io.knotx.repository.fs.FilesystemRepositoryConnectorVerticle",
-    "splitter=io.knotx.knot.splitter.FragmentSplitterVerticle",
+    "splitter=io.knotx.splitter.FragmentSplitterVerticle",
     "assembler=io.knotx.knot.assembler.FragmentAssemblerVerticle",
     "hbsKnot=io.knotx.knot.templating.HandlebarsKnotVerticle",
     "serviceKnot=io.knotx.knot.service.ServiceKnotVerticle",
@@ -102,7 +102,7 @@ For instance, if you want to modify configuration of KnotxServer module, you can
     "server=io.knotx.server.KnotxServerVerticle",
     "httpRepo=io.knotx.repository.http.HttpRepositoryConnectorVerticle",
     "fsRepo=io.knotx.repository.fs.FilesystemRepositoryConnectorVerticle",
-    "splitter=io.knotx.knot.splitter.FragmentSplitterVerticle",
+    "splitter=io.knotx.splitter.FragmentSplitterVerticle",
     "assembler=io.knotx.knot.assembler.FragmentAssemblerVerticle",
     "hbsKnot=io.knotx.knot.templating.HandlebarsKnotVerticle",
     "serviceKnot=io.knotx.knot.service.ServiceKnotVerticle",
