+++ knotx-example/knotx-example-app/config/application.conf
@@ -2,7 +2,7 @@ modules = [
   "server=io.knotx.server.KnotxServerVerticle"
   "httpRepo=io.knotx.repository.http.HttpRepositoryConnectorVerticle"
   "fsRepo=io.knotx.repository.fs.FilesystemRepositoryConnectorVerticle"
-  "splitter=io.knotx.knot.splitter.FragmentSplitterVerticle"
+  "splitter=io.knotx.splitter.FragmentSplitterVerticle"
   "assembler=io.knotx.knot.assembler.FragmentAssemblerVerticle"
   "hbsKnot=io.knotx.knot.templating.HandlebarsKnotVerticle"
   "serviceKnot=io.knotx.knot.service.ServiceKnotVerticle"
