--- knotx-knot/knotx-knot-fragment-splitter/src/main/resources/codegen.json
@@ -1,9 +1,0 @@
-{
-  "name": "knotx-cheatsheet",
-  "generators": [{
-    "kind": "dataObject",
-    "incremental": true,
-    "filename": "'../../../../documentation/.dataobjects/fragment-splitter-dataobjects.adoc'",
-    "templateFilename": "knotx-codegen/dataObjectCheatsheet.templ"
-  }]
-}
