--- knotx-knot/knotx-knot-fragment-splitter/src/test/resources/knotx-fragment-splitter-test.json
@@ -1,5 +1,0 @@
-{
-  "modules": [
-    "splitter=io.knotx.knot.splitter.FragmentSplitterVerticle"
-  ]
-}
