+++ tmp-core/pom.xml
@@ -110,11 +110,6 @@
     </dependency>
     <dependency>
       <groupId>io.knotx</groupId>
-      <artifactId>knotx-knot-fragment-splitter</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>io.knotx</groupId>
       <artifactId>knotx-gateway</artifactId>
       <version>${project.version}</version>
     </dependency>
