+++ tmp-core/src/test/java/io/knotx/splitter/HtmlFragmentSplitterTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.knotx.knot.splitter.impl;
+package io.knotx.splitter;
 
 
 import static org.hamcrest.CoreMatchers.equalTo;
@@ -21,7 +21,6 @@ import static org.hamcrest.MatcherAssert.assertThat;
 
 import io.knotx.dataobjects.Fragment;
 import io.knotx.junit.util.FileReader;
-import io.knotx.knot.splitter.impl.HtmlFragmentSplitter;
 import java.util.List;
 import java.util.stream.IntStream;
 import org.junit.Rule;
