+++ tmp-core/src/test/resources/fragment1.txt
