+++ tmp-core/src/test/resources/fragment2-custom-snippet.txt
