+++ tmp-core/src/test/resources/fragment2-snippet.txt
