+++ tmp-core/src/test/resources/fragment4-custom-snippet.txt
