+++ tmp-core/src/test/resources/fragment4-snippet.txt
