+++ tmp-core/src/test/resources/fragment5-custom-snippet.txt
