+++ tmp-core/src/test/resources/fragment5-snippet.txt
