+++ tmp-core/src/test/resources/fragment6.txt
