+++ tmp-core/src/test/resources/fragment7-custom-snippet.txt
