+++ tmp-core/src/test/resources/fragment7-snippet.txt
