+++ tmp-core/src/test/resources/fragment8.txt
