+++ tmp-core/src/test/resources/fragment9-custom-snippet.txt
