+++ tmp-core/src/test/resources/fragment9-snippet.txt
