+++ tmp-core/src/test/resources/knotx-fragment-splitter-test.json
@@ -1,0 +1,5 @@
+{
+  "modules": [
+    "splitter=io.knotx.splitter.FragmentSplitterVerticle"
+  ]
+}
