+++ tmp-core/src/test/resources/test-many-fragments-custom-snippet.html
