+++ tmp-core/src/test/resources/test-no-fragments.html
