+++ tmp-core/src/test/resources/test-one-fragment-begin.html
