+++ tmp-core/src/test/resources/test-one-fragment-middle.html
