+++ tmp-core/src/test/resources/test-one-fragment-whole.html
